derived removal pt 1
