+++ validator/accounts/wallet_edit.go
@@ -24,8 +24,6 @@ func EditWalletConfigurationCli(cliCtx *cli.Context) error {
 	switch w.KeymanagerKind() {
 	case keymanager.Local:
 		return errors.New("not possible to edit local keymanager configuration")
-	case keymanager.Derived:
-		return errors.New("derived keymanager is not yet supported")
 	case keymanager.Remote:
 		enc, err := w.ReadKeymanagerConfigFromDisk(cliCtx.Context)
 		if err != nil {
