+++ validator/keymanager/BUILD.bazel
@@ -26,7 +26,6 @@ go_test(
     srcs = ["types_test.go"],
     deps = [
         ":go_default_library",
-        "//validator/keymanager/derived:go_default_library",
         "//validator/keymanager/local:go_default_library",
         "//validator/keymanager/remote:go_default_library",
     ],
