--- validator/keymanager/derived/BUILD.bazel
@@ -1,54 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "keymanager.go",
-        "log.go",
-        "mnemonic.go",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/validator/keymanager/derived",
-    visibility = [
-        "//tools:__subpackages__",
-        "//validator:__subpackages__",
-    ],
-    deps = [
-        "//async/event:go_default_library",
-        "//config/fieldparams:go_default_library",
-        "//crypto/bls:go_default_library",
-        "//crypto/rand:go_default_library",
-        "//io/prompt:go_default_library",
-        "//proto/eth/service:go_default_library",
-        "//proto/prysm/v1alpha1/validator-client:go_default_library",
-        "//validator/accounts/iface:go_default_library",
-        "//validator/keymanager:go_default_library",
-        "//validator/keymanager/local:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_tyler_smith_go_bip39//:go_default_library",
-        "@com_github_wealdtech_go_eth2_util//:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = [
-        "eip_test.go",
-        "keymanager_test.go",
-        "mnemonic_test.go",
-    ],
-    embed = [":go_default_library"],
-    deps = [
-        "//config/fieldparams:go_default_library",
-        "//crypto/bls:go_default_library",
-        "//crypto/rand:go_default_library",
-        "//proto/prysm/v1alpha1/validator-client:go_default_library",
-        "//testing/assert:go_default_library",
-        "//testing/require:go_default_library",
-        "//validator/accounts/testing:go_default_library",
-        "//validator/testing:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_tyler_smith_go_bip39//:go_default_library",
-        "@com_github_wealdtech_go_eth2_util//:go_default_library",
-    ],
-)
