+++ validator/keymanager/local/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "import.go",
         "keymanager.go",
         "log.go",
+        "mnemonic.go",
         "refresh.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/validator/keymanager/local",
@@ -23,8 +24,10 @@ go_library(
         "//config/features:go_default_library",
         "//config/fieldparams:go_default_library",
         "//crypto/bls:go_default_library",
+        "//crypto/rand:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//io/file:go_default_library",
+        "//io/prompt:go_default_library",
         "//proto/eth/service:go_default_library",
         "//proto/prysm/v1alpha1/validator-client:go_default_library",
         "//runtime/interop:go_default_library",
@@ -35,8 +38,10 @@ go_library(
         "@com_github_google_uuid//:go_default_library",
         "@com_github_k0kubun_go_ansi//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
+        "@com_github_tyler_smith_go_bip39//:go_default_library",
         "@com_github_schollz_progressbar_v3//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
+        "@com_github_wealdtech_go_eth2_util//:go_default_library",
         "@com_github_wealdtech_go_eth2_wallet_encryptor_keystorev4//:go_default_library",
         "@io_opencensus_go//trace:go_default_library",
     ],
@@ -47,8 +52,10 @@ go_test(
     srcs = [
         "backup_test.go",
         "delete_test.go",
+        "eip_test.go",
         "import_test.go",
         "keymanager_test.go",
+        "mnemonic_test.go",
         "refresh_test.go",
     ],
     embed = [":go_default_library"],
@@ -56,6 +63,7 @@ go_test(
         "//async/event:go_default_library",
         "//config/fieldparams:go_default_library",
         "//crypto/bls:go_default_library",
+        "//crypto/rand:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//proto/eth/service:go_default_library",
         "//proto/prysm/v1alpha1/validator-client:go_default_library",
@@ -63,8 +71,12 @@ go_test(
         "//testing/require:go_default_library",
         "//validator/accounts/testing:go_default_library",
         "//validator/keymanager:go_default_library",
+        "//validator/testing:go_default_library",
         "@com_github_google_uuid//:go_default_library",
+        "@com_github_pkg_errors//:go_default_library",
+        "@com_github_tyler_smith_go_bip39//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
+        "@com_github_wealdtech_go_eth2_util//:go_default_library",
         "@com_github_wealdtech_go_eth2_wallet_encryptor_keystorev4//:go_default_library",
     ],
 )
