+++ validator/keymanager/local/keymanager.go
@@ -18,6 +18,7 @@ import (
 	"github.com/prysmaticlabs/prysm/validator/accounts/iface"
 	"github.com/prysmaticlabs/prysm/validator/accounts/petnames"
 	"github.com/prysmaticlabs/prysm/validator/keymanager"
+	util "github.com/wealdtech/go-eth2-util"
 	keystorev4 "github.com/wealdtech/go-eth2-wallet-encryptor-keystorev4"
 	"go.opencensus.io/trace"
 )
@@ -35,11 +36,18 @@ const (
 	AccountsPath = "accounts"
 	// AccountsKeystoreFileName exposes the name of the keystore file.
 	AccountsKeystoreFileName = "all-accounts.keystore.json"
+	// DerivationPathFormat describes the structure of how keys are derived from a master key.
+	DerivationPathFormat = "m / purpose / coin_type / account_index / withdrawal_key / validating_key"
+	// ValidatingKeyDerivationPathTemplate defining the hierarchical path for validating
+	// keys for Prysm Ethereum validators. According to EIP-2334, the format is as follows:
+	// m / purpose / coin_type / account_index / withdrawal_key / validating_key
+	ValidatingKeyDerivationPathTemplate = "m/12381/3600/%d/0/0"
 )
 
 // Keymanager implementation for local keystores utilizing EIP-2335.
 type Keymanager struct {
 	wallet              iface.Wallet
+	enableMnemonic      bool
 	accountsStore       *accountStore
 	accountsChangedFeed *event.Feed
 }
@@ -48,6 +56,7 @@ type Keymanager struct {
 // a keymanager, such as passwords, the wallet, and more.
 type SetupConfig struct {
 	Wallet           iface.Wallet
+	EnableMnemonic   bool
 	ListenForChanges bool
 }
 
@@ -79,6 +88,7 @@ func ResetCaches() {
 func NewKeymanager(ctx context.Context, cfg *SetupConfig) (*Keymanager, error) {
 	k := &Keymanager{
 		wallet:              cfg.Wallet,
+		enableMnemonic:      cfg.EnableMnemonic,
 		accountsStore:       &accountStore{},
 		accountsChangedFeed: new(event.Feed),
 	}
@@ -125,6 +135,11 @@ func NewInteropKeymanager(_ context.Context, offset, numValidatorKeys uint64) (*
 	orderedPublicKeys = pubKeys
 	lock.Unlock()
 	return k, nil
+}
+
+// EnableMnemonic is active.
+func (km *Keymanager) EnableMnemonic() bool {
+	return km.enableMnemonic
 }
 
 // SubscribeAccountChanges creates an event subscription for a channel
@@ -316,4 +331,29 @@ func (km *Keymanager) CreateAccountsKeystore(
 		Version: encryptor.Version(),
 		Name:    encryptor.Name(),
 	}, nil
+}
+
+// RecoverAccountsFromMnemonic given a mnemonic phrase, is able to regenerate N accounts
+// from a derived seed, encrypt them according to the EIP-2334 JSON standard, and write them
+// to disk. Then, the mnemonic is never stored nor used by the validator.
+func (km *Keymanager) RecoverAccountsFromMnemonic(
+	ctx context.Context, mnemonic, mnemonicPassphrase string, numAccounts int,
+) error {
+	seed, err := seedFromMnemonic(mnemonic, mnemonicPassphrase)
+	if err != nil {
+		return errors.Wrap(err, "could not initialize new wallet seed file")
+	}
+	privKeys := make([][]byte, numAccounts)
+	pubKeys := make([][]byte, numAccounts)
+	for i := 0; i < numAccounts; i++ {
+		privKey, err := util.PrivateKeyFromSeedAndPath(
+			seed, fmt.Sprintf(ValidatingKeyDerivationPathTemplate, i),
+		)
+		if err != nil {
+			return err
+		}
+		privKeys[i] = privKey.Marshal()
+		pubKeys[i] = privKey.PublicKey().Marshal()
+	}
+	return km.ImportKeypairs(ctx, privKeys, pubKeys)
 }
