+++ validator/keymanager/local/mnemonic.go
@@ -1,4 +1,4 @@
-package derived
+package local
 
 import (
 	"fmt"
