+++ validator/rpc/BUILD.bazel
@@ -43,7 +43,6 @@ go_library(
         "//validator/client:go_default_library",
         "//validator/db:go_default_library",
         "//validator/keymanager:go_default_library",
-        "//validator/keymanager/derived:go_default_library",
         "//validator/keymanager/local:go_default_library",
         "//validator/slashing-protection-history:go_default_library",
         "//validator/slashing-protection-history/format:go_default_library",
@@ -107,7 +106,6 @@ go_test(
         "//validator/client:go_default_library",
         "//validator/db/kv:go_default_library",
         "//validator/keymanager:go_default_library",
-        "//validator/keymanager/derived:go_default_library",
         "//validator/keymanager/local:go_default_library",
         "//validator/slashing-protection-history/format:go_default_library",
         "//validator/testing:go_default_library",
